Move spell checker into its own file, no conditional loading
